+++ src/main/java/Add.java
@@ -1,0 +1,5 @@
+public enum Add {
+    TODO,
+    DEADLINE,
+    EVENT
+}
