+++ src/main/java/AlexException.java
@@ -1,5 +1,6 @@
 public class AlexException extends Exception{
     public AlexException(String message) {
+
         super(message);
     }
