+++ src/main/java/Edit.java
@@ -1,0 +1,5 @@
+public enum Edit {
+    MARK,
+    UNMARK,
+    DELETE
+}
