+++ src/main/java/ExitCommand.java
@@ -1,0 +1,12 @@
+public class ExitCommand extends Command{
+    private String command;
+    public ExitCommand(String command) {
+        this.command = command;
+    }
+
+    @Override
+    public void execute() {
+        Ui.bye();
+        System.exit(0);
+    }
+}
