+++ src/main/java/UnknownCommand.java
@@ -1,0 +1,12 @@
+public class UnknownCommand extends Command{
+    private String command;
+
+    public UnknownCommand (String command) {
+        this.command = command;
+    }
+
+    @Override
+    public void execute() {
+        Ui.printAlertForUnknown();
+    }
+}
