+++ src/main/java/View.java
@@ -1,0 +1,4 @@
+public enum View {
+    LISTALL,
+    LISTONEDAY
+}
